requirements.txt
asgiref==3.4.1
Django==3.2.9
django-environ==0.8.1
django-jalali-date==1.0.0
jdatetime==3.6.4
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
uWSGI==2.0.20
